Corrections to AnalogInput ratings

+++ VM_Transmitter/Math/MathUtil.h
@@ -12,7 +12,7 @@ class MathUtil {
 public:
 
 	static float map(long value, float in_min,
-			float in_max, long out_min, long out_max)	{
+			float in_max, float out_min, float out_max)	{
 	  return (float)(value - in_min) * (out_max - out_min) /
 			  (in_max - in_min) + out_min;
 	}
